Detalles del perfil
Al cargar un nuevo pesaje ahora se agrega al usuario. De esta manera se hace visible en el perfil.
Tambien se modifica la vista para que sea visible el avatar del usuario en el perfil.

+++ Services/IUsuarioService.cs
@@ -17,7 +17,7 @@ namespace FitnessManager.Services
         Task<DateTime> GetFechaRegistroAsync(string userId);
         Task<DateTime> GetFechaNacimientoAsync(string userId);
         Task UpdateUserAsync(Usuario user);
-        Task UpdatePesoAsync(string userId, float nuevoPeso);
+        Task UpdatePesoAsync(string userId, double nuevoPeso);
         Task<int> GetEdadAsync(string userId);
         Task SetEdadAsync(string userId, int edad);
         Task DeleteUserAsync(Usuario user);
